Removing the hardcoded user object, no hash

+++ backend/controllers/cards.js
@@ -1,6 +1,6 @@
-const handleError = require('../utils');
-const { errorTypes } = require('../utils');
-const Card = require('../models/card');
+const handleError = require("../utils");
+const { errorTypes } = require("../utils");
+const Card = require("../models/card");
 
 const getCards = (req, res) => {
   Card.find({})
@@ -21,9 +21,27 @@ const createNewCard = (req, res) => {
 };
 
 const deleteCard = (req, res) => {
-  Card.findByIdAndDelete(req.params.cardId)
-    .orFail()
-    .then((card) => res.send({ data: card }))
+  Card.findById(req.params.cardId)
+    .orFail(() => {
+      const error = new Error("Card not found");
+      error.statusCode = 404;
+      throw error;
+    })
+    .then((card) => {
+      if (!(card.owner.toString() === req.user._id)) {
+        const error = new Error("Action forbidden");
+        error.statusCode = 404;
+        throw error;
+      }
+
+      Card.findByIdAndDelete(req.params.cardId)
+        .orFail(() => {
+          const error = new Error("User can delete only own cards");
+          error.statusCode = 403;
+          throw error;
+        })
+        .then((card) => res.send({ data: card }));
+    })
     .catch((err) => {
       handleError(err, req, res);
     });
@@ -32,7 +50,7 @@ const likeCard = (req, res) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
-    { new: true },
+    { new: true }
   )
     .then((card) => res.send({ data: card }))
     .catch((err) => {
@@ -43,7 +61,7 @@ const dislikeCard = (req, res) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } }, // remove _id from the array
-    { new: true },
+    { new: true }
   )
     .then((card) => res.send({ data: card }))
     .catch((err) => {
